Factor out common methods in utils/import_srg_spreadsheet.py
